+++ src/app/_components/PomodoroComponents/ActiveFocusTasksDialog/index.tsx
@@ -2,9 +2,10 @@ import {cn} from "@/app/_lib/utils/cn";
 import {Dialog, DialogPanel, DialogTitle,} from "@headlessui/react";
 import {glassEffectClasses} from "@/app/_components/common";
 import {FocusTask} from "@prisma/client";
+import Button from "@/app/_components/Button";
 import {IFocusTasksData} from "@/app/_lib/hooks/useFocusTasksData";
 import {filterCompletedTasks, filterTodoTasks} from "@/app/_lib/utils/focusTasksHelpers";
-import ActiveFocusTasksSection from "@/app/_components/pomodoro/ActiveFocusTasksDialog/ActiveFocusTaskSection";
+import ActiveFocusTasksSection from "@/app/_components/PomodoroComponents/ActiveFocusTasksDialog/ActiveFocusTaskSection";
 
 interface IFocusTasksDialogProps {
 	isOpen: boolean
