+++ src/app/_components/TopBarMenus/SettingsDialog/sections/AccountSection.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import {ArrowRightStartOnRectangleIcon, UserIcon} from "@heroicons/react/24/solid";
-import Button from "@/app/_components/inputs/Button";
+import Button from "@/app/_components/Button";
 import {signOut, useSession} from "next-auth/react";
 import {ISettingsMenuSection} from "@/app/_components/TopBarMenus/SettingsDialog/SettingsMenuSections";
 import {User} from "next-auth";
