+++ src/app/page.tsx
@@ -1,7 +1,7 @@
-import ThemeBackground from "@/app/_components/theme/ThemeBackground";
+import ThemeBackground from "@/app/_components/ThemeBackground";
 import {cn} from "@/app/_lib/utils/cn";
 import FocusQuote from "@/app/_components/FocusQuote";
-import PomodoroComponents from "@/app/_components/pomodoro/PomodoroComponents";
+import PomodoroComponents from "@/app/_components/PomodoroComponents";
 import TopBarMenus from "@/app/_components/TopBarMenus";
 import SpotifyPlaylistDialog from "@/app/_components/SpotifyPlaylistDialog";
 import getQueryClient from "@/app/_lib/react-query/getQueryClient";
